+++ tests/auto/qvalueaxis/tst_qvalueaxis.cpp
@@ -121,9 +121,9 @@ void tst_QValueAxis::qvalueaxis()
 void tst_QValueAxis::max_raw_data()
 {
     QTest::addColumn<qreal>("max");
-    QTest::newRow("1.0") << 1.0;
-    QTest::newRow("50.0") << 50.0;
-    QTest::newRow("101.0") << 101.0;
+    QTest::newRow("1.0") << (qreal)1.0;
+    QTest::newRow("50.0") << (qreal)50.0;
+    QTest::newRow("101.0") << (qreal)101.0;
 }
 
 void tst_QValueAxis::max_raw()
